modularización de firebase Config
